+++ code/renderers/html/src/entry-preview-docs.ts
@@ -1,2 +1,21 @@
-export { parameters } from './docs/config';
-export { decorators, argTypesEnhancers } from './docs/config';
+import type { ArgTypesEnhancer, DecoratorFunction } from '@storybook/types';
+import { SourceType, enhanceArgTypes } from '@storybook/docs-tools';
+
+import { sourceDecorator } from './docs/sourceDecorator';
+import type { HtmlRenderer } from './types';
+
+export const decorators: DecoratorFunction<HtmlRenderer>[] = [sourceDecorator];
+
+export const parameters = {
+  docs: {
+    story: { inline: true },
+    source: {
+      type: SourceType.DYNAMIC,
+      language: 'html',
+      code: undefined,
+      excludeDecorators: undefined,
+    },
+  },
+};
+
+export const argTypesEnhancers: ArgTypesEnhancer[] = [enhanceArgTypes];
