--- code/renderers/preact/src/docs/config.ts
@@ -1,5 +1,0 @@
-export const parameters = {
-  docs: {
-    story: { inline: true },
-  },
-};
