+++ code/renderers/svelte/src/entry-preview-docs.ts
@@ -1,3 +1,18 @@
-export { parameters } from './docs/config';
+import type { ArgTypesEnhancer, DecoratorFunction } from '@storybook/types';
+import { enhanceArgTypes } from '@storybook/docs-tools';
+import { extractArgTypes } from './docs/extractArgTypes';
+import { extractComponentDescription } from './docs/extractComponentDescription';
+import { sourceDecorator } from './docs/sourceDecorator';
+import type { SvelteRenderer } from './types';
 
-export { decorators, argTypesEnhancers } from './docs/config';
+export const parameters = {
+  docs: {
+    story: { inline: true },
+    extractArgTypes,
+    extractComponentDescription,
+  },
+};
+
+export const decorators: DecoratorFunction<SvelteRenderer>[] = [sourceDecorator];
+
+export const argTypesEnhancers: ArgTypesEnhancer<SvelteRenderer>[] = [enhanceArgTypes];
