+++ code/renderers/vue/src/entry-preview-docs.ts
@@ -1,2 +1,17 @@
-export { parameters } from './docs/config';
-export { decorators, argTypesEnhancers } from './docs/config';
+import type { ArgTypesEnhancer, DecoratorFunction } from '@storybook/types';
+import { extractComponentDescription, enhanceArgTypes } from '@storybook/docs-tools';
+import { extractArgTypes } from './docs/extractArgTypes';
+import { sourceDecorator } from './docs/sourceDecorator';
+import type { VueRenderer } from './types';
+
+export const parameters = {
+  docs: {
+    story: { inline: true, iframeHeight: '120px' },
+    extractArgTypes,
+    extractComponentDescription,
+  },
+};
+
+export const decorators: DecoratorFunction<VueRenderer>[] = [sourceDecorator];
+
+export const argTypesEnhancers: ArgTypesEnhancer<VueRenderer>[] = [enhanceArgTypes];
