+++ code/renderers/web-components/src/entry-preview-docs.ts
@@ -1,2 +1,21 @@
-export { parameters } from './docs/config';
-export { decorators, argTypesEnhancers } from './docs/config';
+import type { ArgTypesEnhancer, DecoratorFunction } from '@storybook/types';
+import { SourceType, enhanceArgTypes } from '@storybook/docs-tools';
+import { extractArgTypes, extractComponentDescription } from './docs/custom-elements';
+import { sourceDecorator } from './docs/sourceDecorator';
+import type { WebComponentsRenderer } from './types';
+
+export const decorators: DecoratorFunction<WebComponentsRenderer>[] = [sourceDecorator];
+
+export const parameters = {
+  docs: {
+    extractArgTypes,
+    extractComponentDescription,
+    story: { inline: true },
+    source: {
+      type: SourceType.DYNAMIC,
+      language: 'html',
+    },
+  },
+};
+
+export const argTypesEnhancers: ArgTypesEnhancer<WebComponentsRenderer>[] = [enhanceArgTypes];
